ci: reorganize tests and speed up regression suite
Split test_compare_perf_report.py into three files with clear names:
- conftest.py: shared helpers (compare_cols, normalize_value, format_diff_details)
- test_perf_report_regression.py: report generation regression tests
- test_compare_perf_reports.py: compare tool tests

Also:
- Switch regression test from subprocess to Python API (~4s faster)
- Add orjson to CI install for faster JSON parsing (~8s on large traces)
- Update test_detect_recompute.py to import from conftest
- Update CI workflow with new filenames

No functional changes — same tests, same assertions, same reference data.

Made-with: Cursor
