+++ .github/workflows/regression-tests.yml
@@ -28,7 +28,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install . pytest openpyxl tensorboard-plugin-profile==2.19.0
+          pip install . pytest openpyxl orjson tensorboard-plugin-profile==2.19.0
 
       - name: Run function-based tests
         run: pytest tests/test_subtract_intervals.py tests/test_graph_mode.py tests/test_kernel_launchers.py
@@ -36,8 +36,11 @@ jobs:
       - name: Run rocprof report tests
         run: pytest tests/test_rocprof_perf_report.py
 
-      - name: Run report regression tests
-        run: pytest tests/test_compare_perf_report.py
+      - name: Run perf report regression tests
+        run: pytest tests/test_perf_report_regression.py
+
+      - name: Run compare perf reports tests
+        run: pytest tests/test_compare_perf_reports.py
 
       - name: Run collective analysis tests
         run: pytest tests/test_collective_analysis.py
